moved notes
